Move Contexts.Utils.contextToType to Import
