+++ packages/codec/lib/basic/decode/index.ts
@@ -446,7 +446,7 @@ function* decodeContractAndContext(
         kind: "known" as const,
         address,
         rawAddress,
-        class: Contexts.Utils.contextToType(context)
+        class: Contexts.Import.contextToType(context)
       }
     };
   } else {
@@ -510,7 +510,7 @@ export function decodeInternalFunction(
 ): Format.Values.FunctionInternalResult {
   let deployedPc: number = Conversion.toBN(deployedPcBytes).toNumber();
   let constructorPc: number = Conversion.toBN(constructorPcBytes).toNumber();
-  let context: Format.Types.ContractType = Contexts.Utils.contextToType(
+  let context: Format.Types.ContractType = Contexts.Import.contextToType(
     info.currentContext
   );
   //before anything else: do we even have an internal functions table?
