+++ packages/codec/lib/contexts/index.ts
@@ -1,4 +1,7 @@
 export * from "./types";
 
+import * as Import from "./import";
+export { Import };
+
 import * as Utils from "./utils";
 export { Utils };
