+++ packages/codec/lib/contexts/utils.ts
@@ -145,24 +145,3 @@ export function normalizeContexts(contexts: Contexts): Contexts {
   //finally, return this mess!
   return newContexts;
 }
-
-export function contextToType(context: Context): Format.Types.ContractType {
-  if (context.contractId !== undefined) {
-    return {
-      typeClass: "contract",
-      kind: "native",
-      id: context.contractId.toString(),
-      typeName: context.contractName,
-      contractKind: context.contractKind,
-      payable: context.payable
-    };
-  } else {
-    return {
-      typeClass: "contract",
-      kind: "foreign",
-      typeName: context.contractName,
-      contractKind: context.contractKind,
-      payable: context.payable
-    };
-  }
-}
